Shape change success
Finally fixed both issues!
Wrong shape fix: I realized that I was thinking in terms of rows while changing the columns section. When I changed my thinking, it was much easier to figure out the conditions for when a ship should be put into the enemy ship formation.
Invisible enemy fix: This one was really simple. While looking over the code, I realized the ship count was increasing regardless of whether an enemy ship was being added to the formation or not. This made it so the game thought there were still ships that needed to be destroyed even though the formation was actually empty.